+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## not released
 
+## v1.1.0 (2024-03-09)
+
+- Added a dialog box to set the extraction keyword and tag
+- Added selection to extract at the folder level vs selecting individual notes
+
 ## v1.0.0 (2024-02-08)
 
 - First version